script: Move event DOM interfaces to script/dom/event/
