+++ components/script/dom/mod.rs
@@ -214,11 +214,9 @@ pub(crate) mod abortsignal;
 #[expect(dead_code)]
 pub(crate) mod abstractrange;
 pub(crate) mod activation;
-pub(crate) mod animationevent;
 pub(crate) mod attr;
 pub(crate) mod audio;
 pub(crate) use self::audio::*;
-pub(crate) mod beforeunloadevent;
 pub(crate) mod bindings;
 pub(crate) mod blob;
 #[cfg(feature = "bluetooth")]
@@ -234,10 +232,7 @@ pub(crate) mod characterdata;
 pub(crate) mod client;
 pub(crate) mod clipboard;
 pub(crate) use self::clipboard::*;
-pub(crate) mod closeevent;
-pub(crate) mod commandevent;
 pub(crate) mod comment;
-pub(crate) mod compositionevent;
 pub(crate) mod console;
 pub(crate) mod cookiestore;
 mod create;
@@ -248,7 +243,6 @@ pub(crate) mod cryptokey;
 pub(crate) mod css;
 pub(crate) use self::css::*;
 pub(crate) mod customelementregistry;
-pub(crate) mod customevent;
 pub(crate) mod customstateset;
 pub(crate) mod datatransfer;
 pub(crate) use self::datatransfer::*;
@@ -278,19 +272,13 @@ pub(crate) mod element;
 pub(crate) mod elementinternals;
 pub(crate) mod encoding;
 pub(crate) use self::encoding::*;
-pub(crate) mod errorevent;
 pub(crate) mod event;
-pub(crate) mod eventsource;
-pub(crate) mod eventtarget;
+pub(crate) use self::event::*;
 pub(crate) mod execcommand;
-pub(crate) mod extendableevent;
-pub(crate) mod extendablemessageevent;
 pub(crate) mod fetchlaterresult;
 pub(crate) mod file;
 pub(crate) use self::file::*;
-pub(crate) mod focusevent;
 pub(crate) mod formdata;
-pub(crate) mod formdataevent;
 #[cfg(feature = "gamepad")]
 pub(crate) mod gamepad;
 #[cfg(feature = "gamepad")]
@@ -299,27 +287,22 @@ pub(crate) mod geolocation;
 pub(crate) use self::geolocation::*;
 pub(crate) mod global_scope_script_execution;
 pub(crate) mod globalscope;
-pub(crate) mod hashchangeevent;
 pub(crate) mod headers;
 pub(crate) mod history;
 pub(crate) mod html;
 pub(crate) use self::html::*;
 pub(crate) mod indexeddb;
 pub(crate) use self::indexeddb::*;
-pub(crate) mod inputevent;
 pub(crate) mod intersectionobserver;
 pub(crate) mod intersectionobserverentry;
-pub(crate) mod keyboardevent;
 pub(crate) mod location;
 pub(crate) mod media;
 pub(crate) use self::media::*;
 pub(crate) mod messagechannel;
-pub(crate) mod messageevent;
 #[expect(dead_code)]
 pub(crate) mod messageport;
 pub(crate) mod mimetype;
 pub(crate) mod mimetypearray;
-pub(crate) mod mouseevent;
 pub(crate) mod mutationobserver;
 pub(crate) mod mutationrecord;
 pub(crate) mod namednodemap;
@@ -333,7 +316,6 @@ pub(crate) mod nodeiterator;
 pub(crate) mod nodelist;
 pub(crate) mod notification;
 pub(crate) mod origin;
-pub(crate) mod pagetransitionevent;
 pub(crate) mod paintsize;
 pub(crate) mod paintworkletglobalscope;
 pub(crate) mod performance;
@@ -343,16 +325,11 @@ pub(crate) mod permissionstatus;
 pub(crate) mod pipelineid;
 pub(crate) mod plugin;
 pub(crate) mod pluginarray;
-#[expect(dead_code)]
-pub(crate) mod pointerevent;
-pub(crate) mod popstateevent;
 pub(crate) mod processinginstruction;
 pub(crate) mod processingoptions;
-pub(crate) mod progressevent;
 #[expect(dead_code)]
 pub(crate) mod promise;
 pub(crate) mod promisenativehandler;
-pub(crate) mod promiserejectionevent;
 pub(crate) mod quotaexceedederror;
 pub(crate) mod radionodelist;
 pub(crate) mod range;
@@ -375,10 +352,8 @@ pub(crate) mod servoparser;
 pub(crate) mod shadowroot;
 pub(crate) mod staticrange;
 pub(crate) mod storage;
-pub(crate) mod storageevent;
 pub(crate) mod stream;
 pub(crate) use self::stream::*;
-pub(crate) mod submitevent;
 pub(crate) mod subtlecrypto;
 pub(crate) mod svg;
 pub(crate) use self::svg::*;
@@ -389,15 +364,11 @@ pub(crate) use self::testing::*;
 pub(crate) mod text;
 pub(crate) mod textcontrol;
 pub(crate) mod timeranges;
-pub(crate) mod toggleevent;
 pub(crate) mod touch;
-pub(crate) mod touchevent;
 pub(crate) mod touchlist;
-pub(crate) mod transitionevent;
 pub(crate) mod treewalker;
 pub(crate) mod trustedtypes;
 pub(crate) use self::trustedtypes::*;
-pub(crate) mod uievent;
 pub(crate) mod url;
 pub(crate) use self::url::*;
 pub(crate) mod useractivation;
@@ -426,7 +397,6 @@ pub(crate) mod webrtc;
 pub(crate) use self::webrtc::*;
 pub(crate) mod webvtt;
 pub(crate) use self::webvtt::*;
-pub(crate) mod wheelevent;
 #[expect(dead_code)]
 pub(crate) mod window;
 pub(crate) mod windowproxy;
